hctest/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.2
python-decouple==3.5
validate-docbr==1.10.0
Faker==9.9.0